feat: compat support and vanquishing the error log
This adds native compatability with gigastructures and stellaris evolved without relying on ignored errors. Should now be fully compatible with anything that doesn't break the vanilla files. Also made a bunch of changes to track down the last of the error log. Should now be clean (at least on game creation).